+++ auto-service-loader/src/main/java/com/anymore/auto/ServiceLazy.kt
@@ -1,10 +1,8 @@
 package com.anymore.auto
 
-import java.util.function.Supplier
-
 /**
  * Created by anymore on 2023/3/28.
  */
-internal class ServiceLazy<T>(private val supplier: Supplier<T>) : SingletonServiceSupplier<T>() {
+internal class ServiceLazy<T>(private val supplier: ServiceFactory<T>) : SingletonServiceSupplier<T>() {
     override fun newInstance() = supplier.get()
 }