+++ auto-service-loader/src/main/java/com/anymore/auto/ServiceSupplier.kt
@@ -1,12 +1,10 @@
 package com.anymore.auto
 
-import java.util.function.Supplier
-
 /**
  * Created by anymore on 2022/4/10.
  */
-internal class ServiceSupplier<T>(val alias: String = "", val supplier: Supplier<out T>) :
-    Supplier<T> {
+internal class ServiceSupplier<T>(val alias: String = "", val supplier: ServiceFactory<T>) :
+    ServiceFactory<T> {
 
     override fun get() = supplier.get()
 }